requirements.txt
pandas==2.1.4
openpyxl==3.1.2
mysql-connector-python==8.3.0
python-dotenv==1.0.0
tabulate==0.9.0 